Add keyboard shortcut
URL files can now be downloaded using a keyboard shortcut so you don't
have to move the mouse and click the button to do it.

Only tested this on Mac. Decided on Ctrl-l as the default, which can be
customised at `chrome://extensions/configureCommands`. Wanted something
with only a single modifier key that wasn't going to clash with another
command. The "Ctrl" string on Mac automatically gets converted to
"Apple", so using "MacCtrl" to get the real "Ctrl" key.