manifest.json: Set `open_in_tab: true`
While it doesn't really look good or fit with the browser UI, this
extension was originally designed before there was an integrated options
page in Chrome. To get it working quickly, enable this setting and open
our options page in a new tab so we can actually interact with it.