Convert `browser.tabs.query` to use promise
The Web Extensions API changes the interface for this call, requiring
you to specify which tabs you want to select from a set of filtering
options (I had tried to use `title` to select only my extension page
tab, but that filter seemed not to be supported). The `.query` call
returns a promise, instead of a callback.