Use Web Extensions downloads API
The mechanism I had been using to download the URL text file doesn't
work in my tests in Firefox 55.0.3. To get around this, use the new
'downloads' Web Extensions API to download the file. Quite a bit nicer,
not as hacky of a solution. Here I left the download `<a>` node just
because at the moment I just want to make the minimum amount of changes
to get the extension ported and in a working state.

In order to use the 'downloads' API, we need to ask permission to use it
in the manifest.